requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
Django==6.0.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.15
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==12.1.1
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rest-framework-simplejwt==0.0.2
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
